[Orchestrator] Allow `Server` to handle different commands.
The message body should now contain a mandatory `"command"` field with the
name of a supported command as value. The rest of the json object depends
on the particular command being transmitted and should contain the needed
information.
Currently supported commands:
- `submitJob`: submits a new job to the Orchestrator. Message should also
have a `"job"` field with the actual json job description as before.
- `cleanTask`: deletes all jobs belonging to a given task from both DBs.
Message should also have a `"task"` field with the name of the task to
delete as value.
- `declareTaskDependency`: declare that a task "B" depends on the completion
of another task "A". Not implemented yet.