[Pilot] Fix weird connection timeout bug.
The connection class is written as a RAII type, but due to the
websocket library calls being asynchronous the connection close
reply from the server sometimes was arriving after the destructor
ran. Add a lock as already done in the constructor to ensure the
destructor doesn' exit *before* the connection is closed.