[Pilot] Add capability for tag-based output transfer
We can now switch different output paths based on the pilot tags.
This requires that the job "output" object is an array with each
element having a "tag" and "files" key.
Traditional format for "output" is still valid, and should be used
when the output path does not depend on the pilot tags.